pyproject.toml
[project]
name = "yampit"
version = "0.1.0"
description = "Builds datasets, jumps through FDBs."
readme = "README.md"
authors = [
    { name = "Tobias Kölling", email = "tobias.koelling@mpimet.mpg.de" },
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.0",
    "polytope-client>=0.7.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipython>=8.29.0",
]
